[🐶加强] example更工程化一些，基础表，权限 [🐷变更] 默认关闭redis依赖 [🐵新增] flywave版本支持分隔，便于人眼识别

+++ wings-slardar/src/main/java/pro/fessional/wings/slardar/security/TypeIdI18nUserDetail.java
@@ -23,7 +23,7 @@ public interface TypeIdI18nUserDetail extends UserDetails {
      *
      * @return 类型
      */
-    long getUserType();
+    int getUserType();
 
 
     /**
